requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
exceptiongroup==1.2.2
idna==3.10
iniconfig==2.1.0
load-dotenv==0.1.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.5
python-dotenv==1.1.0
requests==2.32.3
tomli==2.2.1
urllib3==2.4.0
